requirements.txt
streamlit # frontend
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb